Set the amount of jobs dynamically
Instead of setting a static number of jobs for building flang, use a
function that can calculate the amount of jobs is safe to run on a
system based on the amount of memory required per job and on the amount
of processors available on the said system.

Start using 6144 MiB / job as safe value.  While my measurements have
shown this can be decreased, this value is safe enough and already
reduces the amount of time used to build flang to ~1/3 of the previous
time.